requirements.txt
asgiref==3.3.1
Django==3.1.7
django-bootstrap-modal-forms==2.1.0
django-widget-tweaks==1.4.8
pytz==2021.1
sqlparse==0.4.1
